refactor(sgwallet): remove channel_manager
move the functionality of channel_manager into wallet, so that we can use channel directly in
wallet.
